requirements.txt
greenlet==1.1.0
importlib-metadata==4.5.0
numpy==1.20.3
pandas==1.2.4
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.18
typing-extensions==3.10.0.0
zipp==3.4.1
